Don't enable rope if a `.ropeproject` directory doesn't exist in cwd
I don't want to disable rope completely, but with it on, if I forget and
edit a python file in my home directory, vim hangs while rope tries to
index everything in my home!

This should work OK for me, because I nearly always open a single vim
session in the root of whatever project I'm working on, which will most
likely contain the `.ropeproject` directory (once I create it anyway).